management-plane/requirements.txt
grpcio==1.74.0
grpcio-tools==1.74.0